requirements.txt
absl-py==0.12.0
actionlib==1.12.1
addict==2.4.0
angles==1.9.12
antlr4-python3-runtime==4.8
anyio==3.1.0
appdirs==1.4.4
argon2-cffi==20.1.0
argoverse==1.1.0
async-generator==1.10
attrs==21.2.0
autolab-core==1.1.0
autolab-perception==1.0.0
Babel==2.9.1
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
base-local-planner==1.16.7
black==21.12b0
bleach==3.3.0
bondpy==1.8.5
cached-property==1.5.2
cachetools==4.2.2
camera-calibration==1.15.0
camera-calibration-parsers==1.11.13
carla==0.9.11
catkin==0.7.29
catkin-pkg==0.4.23
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cloudpickle==1.6.0
cmake==3.18.4.post1
colorlog==6.6.0
colour==0.1.5
contextvars==2.4
controller-manager==0.18.2
controller-manager-msgs==0.18.2
cv-bridge==1.13.0
cvxopt==1.2.7
cycler==0.10.0
Cython==0.29.23
dataclasses==0.8
decorator @ file:///home/ktietz/src/ci/decorator_1611930055503/work
defusedxml==0.7.1
deprecation==2.1.0
descartes==1.1.0
detectron2==0.4+cu101
diagnostic-analysis==1.9.7
diagnostic-common-diagnostics==1.9.7
diagnostic-updater==1.9.7
dill==0.3.4
distlib==0.3.1
distro==1.5.0
docopt==0.6.2
docutils==0.17.1
dynamic-reconfigure==1.6.3
entrypoints==0.3
ffmpeg-python==0.2.0
filelock==3.0.12
fire==0.4.0
flake8==4.0.1
flatbuffers==2.0
future==0.18.2
fvcore==0.1.3.post20210317
gazebo-plugins==2.8.7
gazebo-ros==2.8.7
gencpp==0.6.5
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.16
gennodejs==2.0.1
genpy==0.6.14
geodesy==0.5.3
gitdb==4.0.7
GitPython==3.1.17
google-auth==1.30.0
google-auth-oauthlib==0.4.4
gps-common==0.3.1
graphsurgeon @ file:///home/znfs/TensorRT-7.2.3.4/graphsurgeon/graphsurgeon-0.4.5-py2.py3-none-any.whl
grasp-nms==1.0.2
graspnetAPI==1.2.10
grpcio==1.37.1
h5py==3.1.0
hydra-core==1.1.0
idna==2.10
image-geometry==1.13.0
imageio==2.9.0
imgviz==1.2.6
immutables==0.15
importlib-metadata==1.7.0
importlib-resources==3.0.0
iniconfig==1.1.1
interactive-markers==1.11.5
iopath==0.1.8
ipykernel @ file:///tmp/build/80754af9/ipykernel_1596206602906/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///tmp/build/80754af9/ipython_1593447367857/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
ipywidgets==7.6.3
jedi==0.17.0
Jinja2==3.0.1
joblib==1.0.1
joint-state-publisher==1.12.15
joint-state-publisher-gui==1.12.15
jsk-recognition-utils==1.2.15
jsk-rviz-plugins==2.1.7
jsk-topic-tools==2.2.11
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1616770841739/work
jupyter-console==6.4.0
jupyter-core @ file:///tmp/build/80754af9/jupyter_core_1612213308682/work
jupyter-packaging==0.10.2
jupyter-server==1.8.0
jupyterlab==3.0.16
jupyterlab-pygments==0.1.2
jupyterlab-server==2.6.0
jupyterlab-widgets==1.0.0
kdl-parser-py==1.13.1
kiwisolver==1.3.1
knn-pytorch==0.1
labelme==4.5.7
lanelet2-python==1.0.1
lap==0.4.0
laser-geometry==1.6.5
llvmlite==0.31.0
lpips==0.1.3
lxml==4.6.3
lyft-dataset-sdk==0.0.8
Mako==1.1.5
Markdown==3.3.4
MarkupSafe==2.0.1
master-discovery-fkie==0.8.12
master-sync-fkie==0.8.12
matplotlib==3.2.2
mccabe==0.6.1
message-filters==1.14.10
mistune==0.8.4
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
-e git+https://github.com/open-mmlab/mmcv.git@1216e5fe7fbff3299d94f5e153cba74ce854f567#egg=mmcv_full
mmdet==2.14.0
mmdet3d==0.17.3
motmetrics==1.1.3
moveit-commander==1.0.6
moveit-ros-planning-interface==1.0.6
moveit-ros-visualization==1.0.6
multiprocess==0.70.12.2
mypy-extensions==0.4.3
nbclassic==0.3.1
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
nmea-navsat-driver==0.5.2
nose==1.3.7
notebook==6.4.0
numba==0.48.0
numpy==1.19.0
nuscenes-devkit==1.1.9
oauthlib==3.1.0
olefile==0.46
omegaconf==2.1.0
onnx==1.10.1
onnx-graphsurgeon @ file:///home/znfs/TensorRT-7.2.3.4/onnx_graphsurgeon/onnx_graphsurgeon-0.2.6-py2.py3-none-any.whl
onnxruntime==1.8.1
open3d==0.14.1
opencv-python==4.1.2.30
packaging==20.9
pandas==1.1.5
pandocfilters==1.4.3
parso @ file:///tmp/build/80754af9/parso_1617223946239/work
pathspec==0.9.0
pcl==0.0.0.post1
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow @ file:///tmp/build/80754af9/pillow_1617386154532/work
platformdirs==2.0.0
plotly==5.5.0
pluggy==1.0.0
plyfile==0.7.4
pointnet2==0.0.0
polars==0.13.1
portalocker==2.3.0
prettytable==2.2.0
prometheus-client==0.11.0
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1616415428029/work
protobuf==3.17.0
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycodestyle==2.8.0
pycparser==2.20
pycryptodome==3.10.1
pycuda==2021.1
pydot==1.4.2
pyflakes==2.4.0
pygame==2.0.1
pyglet==1.5.0
Pygments @ file:///tmp/build/80754af9/pygments_1621606182707/work
pyntcloud==0.1.6
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyquaternion==0.9.9
pyrsistent==0.17.3
pyserial==3.5
pytest==7.0.0rc1
python-dateutil @ file:///home/ktietz/src/ci/python-dateutil_1611928101742/work
python-pcl==0.3.0a1
python-qt-binding==0.4.3
pytools==2021.2.8
pytz==2021.1
PyWavefront==1.3.3
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==20.0.0
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
qtconsole==5.2.2
QtPy==1.9.0
requests==2.25.1
requests-oauthlib==1.3.0
resource-retriever==1.12.6
rosapi==0.11.12
rosbag==1.14.10
rosboost-cfg==1.14.9
rosbridge-library==0.11.12
rosbridge-server==0.11.12
rosclean==1.14.9
roscreate==1.14.9
rosgraph==1.14.10
roslaunch==1.14.10
roslib==1.14.9
roslint==0.11.2
roslz4==1.14.10
rosmake==1.14.9
rosmaster==1.14.10
rosmsg==1.14.10
rosnode==1.14.10
rosparam==1.14.10
rospkg==1.3.0
rospy==1.14.10
rosserial-python==0.8.0
rosservice==1.14.10
rostest==1.14.10
rostopic==1.14.10
rosunit==1.14.9
roswtf==1.14.10
rqt-action==0.4.9
rqt-bag==0.5.0
rqt-bag-plugins==0.5.0
rqt-console==0.4.9
rqt-dep==0.4.9
rqt-graph==0.4.11
rqt-gui==0.5.2
rqt-gui-py==0.5.2
rqt-image-view==0.4.16
rqt-launch==0.4.8
rqt-logger-level==0.4.8
rqt-moveit==0.5.7
rqt-msg==0.4.8
rqt-multiplot==0.0.10
rqt-nav-view==0.5.7
rqt-plot==0.4.9
rqt-pose-view==0.5.8
rqt-publisher==0.4.8
rqt-py-common==0.5.2
rqt-py-console==0.4.8
rqt-reconfigure==0.5.3
rqt-robot-dashboard==0.5.7
rqt-robot-monitor==0.5.13
rqt-robot-steering==0.5.10
rqt-runtime-monitor==0.5.7
rqt-rviz==0.6.0
rqt-service-caller==0.4.8
rqt-shell==0.4.9
rqt-srv==0.4.8
rqt-tf-tree==0.6.0
rqt-top==0.4.8
rqt-topic==0.4.11
rqt-web==0.4.8
rsa==4.7.2
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
rviz==1.13.15
scikit-build==0.11.1
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.1
Send2Trash==1.5.0
sensor-msgs==1.12.7
setproctitle==1.2.2
Shapely==1.8.0
six==1.16.0
sklearn==0.0
smach==2.0.1
smach-ros==2.0.1
smclib==1.8.5
smmap==4.0.0
sniffio==1.2.0
sound-play==0.3.9
srdfdom==0.5.2
tabulate==0.8.9
tenacity==8.0.1
tensorboard==2.3.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.2
tensorrt @ file:///home/znfs/TensorRT-7.2.3.4/python/tensorrt-7.2.3.4-cp36-none-linux_x86_64.whl
termcolor==1.1.0
terminado==0.10.0
terminaltables==3.1.10
testpath==0.5.0
tf==1.12.1
tf-conversions==1.12.1
tf2-geometry-msgs==0.6.5
tf2-kdl==0.6.5
tf2-py==0.6.5
tf2-ros==0.6.5
tf2-sensor-msgs==0.6.5
thop==0.0.31.post2005241907
threadpoolctl==2.1.0
tifffile==2020.9.3
tomli==1.2.3
tomlkit==0.7.2
topic-tools==1.14.10
torch==1.6.0
torchvision==0.7.0
tornado @ file:///tmp/build/80754af9/tornado_1606942266872/work
tqdm==4.60.0
traitlets==4.3.3
transforms3d==0.3.1
trimesh==3.9.36
turtlebot3-teleop==1.2.5
typed-ast==1.5.1
typing-extensions==4.0.0
unique-id==1.0.6
urdfdom-py==0.4.4
urllib3==1.26.4
virtualenv==20.0.27
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
websocket-client==1.0.1
Werkzeug==2.0.0
widgetsnbextension==3.5.1
wiimote==1.14.0
xacro==1.13.9
xrandr-extend==0.2.0
yacs==0.1.8
yapf==0.31.0
zipp==3.1.0
